# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Webwork Messenger",
        "description": "Webwork Messenger API",
        "license": {
            "name": "Webwork Messenger",
            "url": "https://messenger.webwork-tracker.com"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/chat/{team_id}/{receiver_id}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get messages of two users",
                "parameters":[
                    {
                        "name":"team_id",
                        "in":"path",
                        "required": true,
                        "description":"Get messages of two users",
                        "type": "string"
                    },
                    {
                        "name":"receiver_id",
                        "in":"path",
                        "required": true,
                        "description":"Get messages of two users",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Chat"
                        }
                    }
                }
            }
        },
        "/api/chat/send-message": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Send message",
                "parameters":[
                    {
                        "name":"message",
                        "in":"body",
                        "description":"Send message",
                        "schema":{
                            "$ref":"#/definitions/Chat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Chat"
                        }
                    }
                }
            }
        },
        "/api/chat/send-message/channel": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Send message to channel",
                "parameters":[
                    {
                        "name":"message",
                        "in":"body",
                        "description":"Send message to channel",
                        "schema":{
                            "$ref":"#/definitions/Channel_Messages"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Channel_Messages"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Chat": {
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "receiver_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "channel": { 
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Channel_Messages": {
            "required":[
                "message",
                "user_id"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "user_id": {
                    "type": "string"
                },
                "channel_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}